Initial Commit
Added barebone templates (index.php, header.php, footer.php, and 
sidebar.php)
as well as a minimal sass stylesheet based on, but with less clutter 
than
underscores. The main structure follows a simple

 * header
 * main
 * aside
 * footer

pattern. The intention is to make the whole layout run on CSS Grid with 
a
mobile first approach.